+++ tests/integration_tests/functional/test_cmd_line_parameters.py
@@ -4,7 +4,9 @@
 
 import logging
 import platform
+from pathlib import Path
 
+import host_tools.logging as log_tools
 from host_tools.cargo_build import get_firecracker_binaries
 from conftest import _test_images_s3_bucket
 from framework.artifacts import ArtifactCollection
@@ -80,3 +82,111 @@ def test_describe_snapshot_all_versions(bin_cloner_path):
         assert code == 0
         assert stderr == ""
         assert target_version in stdout
+
+
+def test_cli_metrics_path(test_microvm_with_api):
+    """
+    Test --metrics-path parameter
+
+    @type: functional
+    """
+    microvm = test_microvm_with_api
+    metrics_fifo_path = Path(microvm.path) / "metrics_ndjson.fifo"
+    metrics_fifo = log_tools.Fifo(metrics_fifo_path)
+    microvm.spawn(metrics_path=metrics_fifo_path)
+    microvm.basic_config()
+    microvm.start()
+
+    metrics = microvm.flush_metrics(metrics_fifo)
+
+    exp_keys = [
+        "utc_timestamp_ms",
+        "api_server",
+        "balloon",
+        "block",
+        "deprecated_api",
+        "get_api_requests",
+        "i8042",
+        "latencies_us",
+        "logger",
+        "mmds",
+        "net",
+        "patch_api_requests",
+        "put_api_requests",
+        "seccomp",
+        "vcpu",
+        "vmm",
+        "uart",
+        "signals",
+        "vsock",
+    ]
+
+    if platform.machine() == "aarch64":
+        exp_keys.append("rtc")
+
+    assert set(metrics.keys()) == set(exp_keys)
+
+
+def test_cli_metrics_path_if_metrics_initialized_twice_fail(test_microvm_with_api):
+    """
+    Given: a running firecracker with metrics configured with the CLI option
+    When: Configure metrics via API
+    Then: API returns an error
+
+    @type: functional
+    """
+    microvm = test_microvm_with_api
+
+    # First configure the µvm metrics with --metrics-path
+    metrics_path = Path(microvm.path) / "metrics.ndjson"
+    metrics_path.touch()
+    microvm.spawn(metrics_path=metrics_path)
+
+    # Then try to configure it with PUT /metrics
+    metrics2_path = Path(microvm.path) / "metrics2.ndjson"
+    metrics2_path.touch()
+    response = microvm.metrics.put(
+        metrics_path=microvm.create_jailed_resource(metrics2_path)
+    )
+
+    # It should fail with HTTP 400 because it's already configured
+    assert response.status_code == 400
+    assert response.json() == {
+        "fault_message": "Reinitialization of metrics not allowed."
+    }
+
+
+def test_cli_metrics_if_resume_no_metrics(test_microvm_with_api, microvm_factory):
+    """
+    Check that metrics configuration is not part of the snapshot
+
+    @type: functional
+    """
+    # Given: a snapshot of a FC with metrics configured with the CLI option
+    uvm1 = test_microvm_with_api
+    metrics_path = Path(uvm1.path) / "metrics.ndjson"
+    metrics_path.touch()
+    uvm1.spawn(metrics_path=metrics_path)
+    uvm1.basic_config()
+    uvm1.start()
+
+    mem_path = Path(uvm1.jailer.chroot_path()) / "test.mem"
+    snapshot_path = Path(uvm1.jailer.chroot_path()) / "test.snap"
+    uvm1.pause_to_snapshot(
+        mem_file_path=mem_path.name,
+        snapshot_path=snapshot_path.name,
+    )
+    assert mem_path.exists()
+
+    # When: restoring from the snapshot
+    uvm2 = microvm_factory.build()
+    uvm2.spawn()
+    uvm2.restore_from_snapshot(
+        snapshot_vmstate=snapshot_path,
+        snapshot_mem=mem_path,
+        snapshot_disks=[uvm1.rootfs_file],
+    )
+
+    # Then: the old metrics configuration does not exist
+    metrics2 = Path(uvm2.jailer.chroot_path()) / metrics_path.name
+    assert not metrics2.exists()