+++ CHANGELOG.md
@@ -11,6 +11,9 @@
   making it safe to snapshot uVMs running on a newer host CPU (Cascade Lake)
   and restore on a host that has a Skylake CPU.
 
+- Added a new CLI option `--metrics-path PATH`. It accepts a file parameter
+  where metrics will be sent to.
+
 ### Fixed
 
 - Make the `T2` template more robust by explicitly disabling additional